feat: Fetch all available police crime dates
The `policeService` has been updated to fetch the full list of available crime dates from the Police.uk API, rather than just the latest. This allows the application to present a date selection UI to the user.

A fallback mechanism has also been implemented to generate a list of the last 12 months if the API call fails.